refactor(web): consolidate lib docs and sort-by-meta helpers
Fold cohesive lib/ modules together (16 -> 14 files):
- docs/nav.ts + docs/types.ts -> lib/docs.ts (one docs-nav domain)
- sort-by-meta.ts -> lib/llms.ts (its only consumers are the llms.txt routes)

Update docs.config.ts and the CLI docsConfigTemplate() (templates.ts), which both
import DocsConfig from ./src/lib/docs/types, so a scaffolded fork's docs.config.ts
keeps resolving. Audit doc updated with the lib/ consolidation rationale.

No behavior change: tsc, oxlint, oxfmt clean; 85 CLI tests pass.
