+++ web/next/docs.config.ts
@@ -1,6 +1,6 @@
 import { site } from "@packages/config/site"
 
-import type { DocsConfig } from "./src/lib/docs/types"
+import type { DocsConfig } from "./src/lib/docs"
 
 // Single source of truth for docs structure and metadata. Groups are ordered arrays; each item is a single-key record keyed by a page URL (value = metadata) or a subgroup label (value = nested items). Keys are literal URLs ("/docs" = the docs index).
 // title/description are synced into each MDX's frontmatter by the web/next build/dev (.github/scripts/docs.ts); `label` overrides the sidebar label (defaults to title). Group name is independent of the page URL, so pages regroup without changing links.
