+++ web/next/src/app/(console)/console/layout.tsx
@@ -5,7 +5,7 @@ import { SidebarShell } from "@/components/shell/sidebar-shell"
 import { SidebarUserMenu } from "@/components/shell/user-menu"
 import { SidebarMenu } from "@/components/ui/sidebar"
 import { assertConsoleAccess } from "@/lib/auth/console"
-import { resolveDocsNav } from "@/lib/docs/nav"
+import { resolveDocsNav } from "@/lib/docs"
 
 // Force per-request rendering so the access check runs on every request and the console is never statically prerendered/cached for anonymous users.
 export const dynamic = "force-dynamic"
