+++ web/next/src/components/shell/floating-trigger.tsx
@@ -3,7 +3,7 @@
 import { usePathname } from "next/navigation"
 
 import { SidebarTrigger } from "@/components/ui/sidebar"
-import { isDocsPath } from "@/lib/docs/nav"
+import { isDocsPath } from "@/lib/docs"
 import { cn } from "@/lib/utils"
 
 // Shared floating sidebar trigger. In the docs area the sidebar is offcanvas, so on desktop this becomes an edge tab; elsewhere the in-sidebar header trigger covers desktop and this stays hidden. On mobile it's always a labeled button bottom-right. Label: "Docs" in docs, "Menu" elsewhere.
